Disable weight == 0 validation
This was a feature request. Yes, this makes it possible to bring data
into a state that causes endless crashes due to division by zero if the
grade with weight zero is the only one. That could be prevented by a
more elaborate validation preventing this from ever happening, but I
really can't be bothered.
This is a pile of dung and I don't know why I still support it.
I need help.